Move PluginFinder to Nop.Services project
