+++ src/Libraries/Nop.Services/Payments/PaymentService.cs
@@ -5,9 +5,9 @@ using Nop.Core;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Orders;
 using Nop.Core.Domain.Payments;
-using Nop.Core.Plugins;
 using Nop.Services.Catalog;
 using Nop.Services.Configuration;
+using Nop.Services.Plugins;
 
 namespace Nop.Services.Payments
 {
