+++ src/Libraries/Nop.Services/Plugins/PluginFinder.cs
@@ -1,0 +1,219 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Nop.Core.Domain.Customers;
+using Nop.Core.Plugins;
+
+namespace Nop.Services.Plugins
+{
+    /// <summary>
+    /// Plugin finder
+    /// </summary>
+    public class PluginFinder : IPluginFinder
+    {
+        #region Fields
+
+        private IList<PluginDescriptor> _plugins;
+        private bool _arePluginsLoaded;
+
+        #endregion
+
+        #region Utilities
+
+        /// <summary>
+        /// Ensure plugins are loaded
+        /// </summary>
+        protected virtual void EnsurePluginsAreLoaded()
+        {
+            if (!_arePluginsLoaded)
+            {
+                var foundPlugins = PluginManager.ReferencedPlugins.ToList();
+                foundPlugins.Sort();
+                _plugins = foundPlugins.ToList();
+
+                _arePluginsLoaded = true;
+            }
+        }
+
+        /// <summary>
+        /// Check whether the plugin is available in a certain store
+        /// </summary>
+        /// <param name="pluginDescriptor">Plugin descriptor to check</param>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <returns>true - available; false - no</returns>
+        protected virtual bool CheckLoadMode(PluginDescriptor pluginDescriptor, LoadPluginsMode loadMode)
+        {
+            if (pluginDescriptor == null)
+                throw new ArgumentNullException(nameof(pluginDescriptor));
+
+            switch (loadMode)
+            {
+                case LoadPluginsMode.All:
+                    //no filering
+                    return true;
+                case LoadPluginsMode.InstalledOnly:
+                    return pluginDescriptor.Installed;
+                case LoadPluginsMode.NotInstalledOnly:
+                    return !pluginDescriptor.Installed;
+                default:
+                    throw new Exception("Not supported LoadPluginsMode");
+            }
+        }
+
+        /// <summary>
+        /// Check whether the plugin is in a certain group
+        /// </summary>
+        /// <param name="pluginDescriptor">Plugin descriptor to check</param>
+        /// <param name="group">Group</param>
+        /// <returns>true - available; false - no</returns>
+        protected virtual bool CheckGroup(PluginDescriptor pluginDescriptor, string group)
+        {
+            if (pluginDescriptor == null)
+                throw new ArgumentNullException(nameof(pluginDescriptor));
+
+            if (string.IsNullOrEmpty(group))
+                return true;
+
+            return group.Equals(pluginDescriptor.Group, StringComparison.InvariantCultureIgnoreCase);
+        }
+
+        #endregion
+        
+        #region Methods
+
+        /// <summary>
+        /// Check whether the plugin is available in a certain store
+        /// </summary>
+        /// <param name="pluginDescriptor">Plugin descriptor to check</param>
+        /// <param name="storeId">Store identifier to check</param>
+        /// <returns>true - available; false - no</returns>
+        public virtual bool AuthenticateStore(PluginDescriptor pluginDescriptor, int storeId)
+        {
+            if (pluginDescriptor == null)
+                throw new ArgumentNullException(nameof(pluginDescriptor));
+
+            //no validation required
+            if (storeId == 0)
+                return true;
+
+            if (!pluginDescriptor.LimitedToStores.Any())
+                return true;
+
+            return pluginDescriptor.LimitedToStores.Contains(storeId);
+        }
+
+        /// <summary>
+        /// Check that plugin is available for the specified customer
+        /// </summary>
+        /// <param name="pluginDescriptor">Plugin descriptor to check</param>
+        /// <param name="customer">Customer</param>
+        /// <returns>True if authorized; otherwise, false</returns>
+        public virtual bool AuthorizedForUser(PluginDescriptor pluginDescriptor, Customer customer)
+        {
+            if (pluginDescriptor == null)
+                throw new ArgumentNullException(nameof(pluginDescriptor));
+
+            if (customer == null || !pluginDescriptor.LimitedToCustomerRoles.Any())
+                return true;
+
+            var customerRoleIds = customer.CustomerRoles.Where(role => role.Active).Select(role => role.Id);
+
+            return pluginDescriptor.LimitedToCustomerRoles.Intersect(customerRoleIds).Any();
+        }
+
+        /// <summary>
+        /// Gets plugin groups
+        /// </summary>
+        /// <returns>Plugins groups</returns>
+        public virtual IEnumerable<string> GetPluginGroups()
+        {
+            return GetPluginDescriptors(LoadPluginsMode.All).Select(x => x.Group).Distinct().OrderBy(x => x);
+        }
+
+        /// <summary>
+        /// Gets plugins
+        /// </summary>
+        /// <typeparam name="T">The type of plugins to get.</typeparam>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <param name="customer">Load records allowed only to a specified customer; pass null to ignore ACL permissions</param>
+        /// <param name="storeId">Load records allowed only in a specified store; pass 0 to load all records</param>
+        /// <param name="group">Filter by plugin group; pass null to load all records</param>
+        /// <returns>Plugins</returns>
+        public virtual IEnumerable<T> GetPlugins<T>(LoadPluginsMode loadMode = LoadPluginsMode.InstalledOnly,
+            Customer customer = null, int storeId = 0, string group = null) where T : class, IPlugin
+        {
+            return GetPluginDescriptors<T>(loadMode, customer, storeId, group).Select(p => p.Instance<T>());
+        }
+
+        /// <summary>
+        /// Get plugin descriptors
+        /// </summary>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <param name="customer">Load records allowed only to a specified customer; pass null to ignore ACL permissions</param>
+        /// <param name="storeId">Load records allowed only in a specified store; pass 0 to load all records</param>
+        /// <param name="group">Filter by plugin group; pass null to load all records</param>
+        /// <returns>Plugin descriptors</returns>
+        public virtual IEnumerable<PluginDescriptor> GetPluginDescriptors(LoadPluginsMode loadMode = LoadPluginsMode.InstalledOnly,
+            Customer customer = null, int storeId = 0, string group = null)
+        {
+            //ensure plugins are loaded
+            EnsurePluginsAreLoaded();
+
+            return _plugins.Where(p => CheckLoadMode(p, loadMode) && AuthorizedForUser(p, customer) && AuthenticateStore(p, storeId) && CheckGroup(p, group));
+        }
+
+        /// <summary>
+        /// Get plugin descriptors
+        /// </summary>
+        /// <typeparam name="T">The type of plugin to get.</typeparam>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <param name="customer">Load records allowed only to a specified customer; pass null to ignore ACL permissions</param>
+        /// <param name="storeId">Load records allowed only in a specified store; pass 0 to load all records</param>
+        /// <param name="group">Filter by plugin group; pass null to load all records</param>
+        /// <returns>Plugin descriptors</returns>
+        public virtual IEnumerable<PluginDescriptor> GetPluginDescriptors<T>(LoadPluginsMode loadMode = LoadPluginsMode.InstalledOnly,
+            Customer customer = null, int storeId = 0, string group = null) 
+            where T : class, IPlugin
+        {
+            return GetPluginDescriptors(loadMode, customer, storeId, group)
+                .Where(p => typeof(T).IsAssignableFrom(p.PluginType));
+        }
+
+        /// <summary>
+        /// Get a plugin descriptor by its system name
+        /// </summary>
+        /// <param name="systemName">Plugin system name</param>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <returns>>Plugin descriptor</returns>
+        public virtual PluginDescriptor GetPluginDescriptorBySystemName(string systemName, LoadPluginsMode loadMode = LoadPluginsMode.InstalledOnly)
+        {
+            return GetPluginDescriptors(loadMode)
+                .SingleOrDefault(p => p.SystemName.Equals(systemName, StringComparison.InvariantCultureIgnoreCase));
+        }
+
+        /// <summary>
+        /// Get a plugin descriptor by its system name
+        /// </summary>
+        /// <typeparam name="T">The type of plugin to get.</typeparam>
+        /// <param name="systemName">Plugin system name</param>
+        /// <param name="loadMode">Load plugins mode</param>
+        /// <returns>>Plugin descriptor</returns>
+        public virtual PluginDescriptor GetPluginDescriptorBySystemName<T>(string systemName, LoadPluginsMode loadMode = LoadPluginsMode.InstalledOnly)
+            where T : class, IPlugin
+        {
+            return GetPluginDescriptors<T>(loadMode)
+                .SingleOrDefault(p => p.SystemName.Equals(systemName, StringComparison.InvariantCultureIgnoreCase));
+        }
+
+        /// <summary>
+        /// Reload plugins
+        /// </summary>
+        public virtual void ReloadPlugins()
+        {
+            _arePluginsLoaded = false;
+            EnsurePluginsAreLoaded();
+        }
+
+        #endregion
+    }
+}
