+++ src/Presentation/Nop.Web/Areas/Admin/Views/Shared/Menu.cshtml
@@ -1,4 +1,4 @@
-﻿@using Nop.Core.Plugins
+﻿@using Nop.Services.Plugins
 @using Nop.Services.Security
 @using Nop.Web.Framework.Menu
 @inject IPluginFinder pluginFinder
