+++ src/Presentation/Nop.Web/Areas/Admin/Views/Shared/_ConfigurePlugin.cshtml
@@ -5,6 +5,7 @@
 @using Nop.Services.Cms
 @using Nop.Services.Common
 @using Nop.Services.Payments
+@using Nop.Services.Plugins
 @using Nop.Services.Shipping
 @using Nop.Services.Shipping.Pickup
 @using Nop.Services.Tax
