+++ src/Tests/Nop.Services.Tests/Discounts/DiscountServiceTests.cs
@@ -6,11 +6,11 @@ using Nop.Core.Caching;
 using Nop.Core.Data;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Discounts;
-using Nop.Core.Plugins;
 using Nop.Services.Catalog;
 using Nop.Services.Discounts;
 using Nop.Services.Events;
 using Nop.Services.Localization;
+using Nop.Services.Plugins;
 using Nop.Tests;
 using NUnit.Framework;
 using Rhino.Mocks;
