+++ src/Tests/Nop.Services.Tests/Payments/PaymentServiceTests.cs
@@ -2,9 +2,9 @@
 using System.Linq;
 using Nop.Core.Domain.Orders;
 using Nop.Core.Domain.Payments;
-using Nop.Core.Plugins;
 using Nop.Services.Configuration;
 using Nop.Services.Payments;
+using Nop.Services.Plugins;
 using Nop.Tests;
 using NUnit.Framework;
 using Rhino.Mocks;
